+++ local/lib/python3/cmk_addons/plugins/oposs_pbs/server_side_calls/oposs_pbs.py
@@ -18,7 +18,7 @@ class Params(BaseModel):
     task_limit: int = 1000
     timeout: int | None = None
     refresh_budget: int | None = None
-    piggyback_template: str = "{id}"
+    piggyback_template: str = "{guest}"
     piggyback_regex: str | None = None
     no_piggyback: list[str] = []